🚚 Rename: 컴포넌트 폴더 생성 및 구조 변경
