+++ src/pages/shared.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import Header from '../components/Header';
-import Nav from '../components/Nav';
-import Cardlist from '../components/Cardlist';
-import Footer from '../components/Footer';
+import Header from '../components/common/Header';
+import Nav from '../components/shared/Nav';
+import Cardlist from '../components/shared/Cardlist';
+import Footer from '../components/common/Footer';
 
 const Shared = () => {
   return (
